fetching certificate directly but not via application flow
