+++ samples/corda/corda-simple-application/workflows-kotlin/src/main/kotlin/com/example/flow/AssetFlow.kt
@@ -771,33 +771,6 @@ class GetSimpleAssetStateAndContractId(
 }
 
 /**
- * The GetOurCertificateBase64 flow is used to fetch the certificate of the party owning the node in base64 format.
- */
-@InitiatingFlow
-@StartableByRPC
-class GetOurCertificateBase64() : FlowLogic<String>() {
-    @Suspendable
-    override fun call(): String {
-
-        return Base64.getEncoder().encodeToString(ourIdentityAndCert.certificate.toString().toByteArray())
-    }
-}
-
-/**
- * The GetOurIdentity flow is used to fetch the identity of the party owning the node.
- */
-@InitiatingFlow
-@StartableByRPC
-class GetOurIdentity() : FlowLogic<String>() {
-    @Suspendable
-    override fun call(): String {
-
-        println("ourIdentity: ${ourIdentity}")
-        return ourIdentity.toString()
-    }
-}
-
-/**
  * The marshalFungibleAsset function is used to obtain the JSON encoding of the fungible asset of interest to the user.
  * This function is typically called by the application client which may not know the full details of the asset.
  *
